BIG UPDATE: Read description
This might not be functional because I'm having issues actually running the code, something about the .SystemManager is messing stuff up so I'm coding in the dark.

Altered each time entry to be an  array [thermostat activity, temperature]. I think I managed to change every instance that was attempting to get just the temperature to accessing the first index of this new array.

Also started setting up the bones for tracking energy in the ThermostatAnalyzer.